+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.0.2
 
-* A `Map` can have several configurations;
-* Scores are retrieved from configurations resource.
+* A `Map` can have several routes;
+* Scores are retrieved from routes resource;
+* Information on the route selector should be more readable now.
 
 ## 0.0.1